src/requirements.txt
openai
PyPDF2==2.12.1
scikit-learn
numpy
tiktoken
pandas
pinecone-client

#tomli
